posts/creating-a-personal-blog-part-3-adding-contact-functionality-and-google-recaptcha-v3-to-django/source code/Blog/requirements.txt
Django==3.0.4
django-recaptcha==2.0.6